geotiff: extract _sources.py (#2228)
Mechanical move of the read-side I/O layer out of _reader.py into a
new _sources.py. No behaviour change, no public API change.

Moved:
- local mmap source (_MmapCache, _FileSource)
- HTTP source with SSRF allow-list and DNS-rebind IP pinning
  (_HTTPSource, _validate_http_url, _ip_is_private, UnsafeURLError,
  pinned-pool factories)
- fsspec cloud source (_CloudSource, _resolve_max_cloud_bytes,
  CloudSizeLimitError)
- BytesIO source (_BytesIOSource)
- byte-range coalescing (coalesce_ranges, split_coalesced_bytes,
  COALESCE_GAP_THRESHOLD_DEFAULT)
- source dispatch (_open_source, _is_fsspec_uri, _is_file_like,
  _coerce_path)
- per-tile byte budget (MAX_TILE_BYTES_DEFAULT, _max_tile_bytes_from_env)

_reader.py re-imports everything from _sources at the top so the
in-module call sites keep their bare names and the historical
import surface used by _sidecar, _vrt, the backends, and the test
suite (which monkeypatches names like _reader._HTTPSource) stays
intact. _reader.py drops from 3453 to 2097 lines.

Closes #2228. Part of #2211.
